Разбил на модули и добавил сборку webpack
