+++ .eslintrc.yml
@@ -2,5 +2,10 @@ env:
   es6: true
   browser: true
   commonjs: true
+  node: true
 
 extends: 'htmlacademy/es6'
+
+parserOptions:
+  ecmaVersion: 2015
+  sourceType: 'module'
