+++ webpack.config.js
@@ -1,0 +1,17 @@
+const path = require(`path`);
+module.exports = {
+  mode: `development`,
+  entry: `./src/main.js`,
+  output: {
+    filename: `bundle.js`,
+    path: path.join(__dirname, `public`)
+  },
+  devtool: `source-map`,
+  devServer: {
+    contentBase: path.join(__dirname, `public`), // Где искать сборку
+    publicPath: `http://localhost:8080/`, // Веб адрес сборки
+    compress: true, // Сжатие
+    overlay: true, // Выводит ошибки на экране.
+    watchContentBase: true
+  }
+};
